refactor(match):아이콘 및 match 페이지 수정
